+++ android/src/main/kotlin/hallo/folder_file_saver/FolderFileSaverPlugin.kt
@@ -1,6 +1,5 @@
 package hallo.folder_file_saver
 
-import android.Manifest
 import android.content.Intent
 import android.content.pm.ApplicationInfo
 import android.content.pm.PackageManager
@@ -9,15 +8,21 @@ import android.graphics.BitmapFactory
 import android.net.Uri
 import android.os.Environment
 import android.provider.Settings
-import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
 
+import android.app.Activity
+import androidx.annotation.NonNull
+import android.content.Context          
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+import io.flutter.embedding.engine.plugins.activity.ActivityAware
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
-import io.flutter.plugin.common.PluginRegistry
-import io.flutter.plugin.common.PluginRegistry.Registrar
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
+//import io.flutter.plugin.common.PluginRegistry.Registrar
+//import io.flutter.plugin.common.PluginRegistry.RequestPermissionsResultListener
 
 import java.io.File
 import java.io.FileOutputStream
@@ -25,51 +30,70 @@ import java.io.IOException
 
 const val CHANNEL_NAME = "folder_file_saver"
 
-class FolderFileSaverPlugin(private val registrar: Registrar) : MethodCallHandler, PluginRegistry.RequestPermissionsResultListener {
+class FolderFileSaverPlugin: FlutterPlugin, MethodCallHandler, ActivityAware {
+    
+    private lateinit var channel : MethodChannel
+    private lateinit var context: Context
+    private lateinit var activity: Activity
 
-    private val permission = Manifest.permission.WRITE_EXTERNAL_STORAGE
-    private val mChannel = MethodChannel(registrar.messenger(), CHANNEL_NAME)
-
-    companion object {
-        @JvmStatic
-        fun registerWith(registrar: Registrar) {
-            val channel = MethodChannel(registrar.messenger(), CHANNEL_NAME)
-            val folderFileSaverPlugin: FolderFileSaverPlugin = FolderFileSaverPlugin(registrar)
-            channel.setMethodCallHandler(folderFileSaverPlugin)
-            registrar.addRequestPermissionsResultListener(folderFileSaverPlugin)
-        }
+    override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
+        channel = MethodChannel(flutterPluginBinding.binaryMessenger, CHANNEL_NAME)
+        channel.setMethodCallHandler(this)
+        context = flutterPluginBinding.applicationContext
     }
 
-    override fun onMethodCall(call: MethodCall, result: Result) {
-        when (call.method) {
-            "saveFileToFolderExt" -> {
-                val filePath = call.argument<String>("filePath")!!
-                result.success(saveFileToFolderExt(filePath))
-            }
-            "saveImage" -> {
-                val imagePath = call.argument<String>("pathImage")!!
-                val width: Int = call.argument<Int>("width")!!
-                val height: Int = call.argument<Int>("height")!!
-                if (width == 0 && height == 0) {
-                    result.success(saveFileToFolderExt(imagePath))
-                } else {
-                    try {
-                        result.success(saveFileToFolderExt(resizeTo(imagePath, width, height)))
-                    } catch (e: IOException) {
-                        e.printStackTrace()
+    override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
+        try {
+            when (call.method) {
+                "saveFileToFolderExt" -> {
+                    val filePath = call.argument<String>("filePath")!!
+                    val removeOriginFile: Boolean = call.argument<Boolean>("removeOriginFile")!!
+                    result.success(saveFileToFolderExt(filePath,removeOriginFile))
+                }
+                "saveImage" -> {
+                    val imagePath = call.argument<String>("pathImage")!!
+                    val width: Int = call.argument<Int>("width")!!
+                    val height: Int = call.argument<Int>("height")!!
+                    val removeOriginFile: Boolean = call.argument<Boolean>("removeOriginFile")!!
+                    if (width == 0 && height == 0) {
+                        result.success(saveFileToFolderExt(imagePath,removeOriginFile))
+                    } else {
+                        val resultPath = saveFileToFolderExt(resizeTo(imagePath, width, height),removeOriginFile)
+                        result.success(resultPath)
                     }
                 }
+                "openSetting" -> {
+                    openSettingsPermission()
+                    result.success(true)
+                }
+                else -> result.notImplemented()
             }
-            "requestPermission" -> {
-                requirePermission()
-                result.success(true)
-            }
-            "openSetting" -> {
-                openSettingsPermission()
-                result.success(true)
-            }
-            else -> result.notImplemented()
+        } catch (e: IOException) {
+            e.printStackTrace()
         }
+        
+    }
+
+    override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
+        channel.setMethodCallHandler(null)
+    }
+
+    // Implements Activity Aware
+    
+    override fun onDetachedFromActivity() {
+        // TODO("Not yet implemented")
+    }
+
+    override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
+        // TODO("Not yet implemented")
+    }
+
+    override fun onAttachedToActivity(binding: ActivityPluginBinding) {
+        activity = binding.activity;
+    }
+
+    override fun onDetachedFromActivityForConfigChanges() {
+        // TODO("Not yet implemented")
     }
 
     private fun resizeTo(pathImage: String, width: Int, height: Int): String {
@@ -90,16 +114,16 @@ class FolderFileSaverPlugin(private val registrar: Registrar) : MethodCallHandle
         return file.path
     }
 
-    private fun saveFileToFolderExt(filePath: String = ""): String {
-        val ctx = registrar.activeContext().applicationContext
+    private fun saveFileToFolderExt(filePath: String = "", removeOriginFile:Boolean = true): String {
         return try {
             val originalFile = File(filePath)
             val file = createFolderOfFile(originalFile.extension)
             originalFile.copyTo(file)
-
+            if(removeOriginFile){
+                originalFile.delete();
+            }
             val uri = Uri.fromFile(file)
-
-            ctx.sendBroadcast(Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, uri))
+            context.sendBroadcast(Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, uri))
             return uri.toString()
         } catch (e: IOException) {
             e.printStackTrace()
@@ -115,7 +139,10 @@ class FolderFileSaverPlugin(private val registrar: Registrar) : MethodCallHandle
             "m4a" -> "Audios"
             else -> "Documents"
         }
-        val storePath = Environment.getExternalStorageDirectory().absolutePath + File.separator + appNamed() + "/" + appNamed() + " " + folderExtNamed
+        val file = ContextCompat.getExternalFilesDirs(context, null)
+        val isNull = file == null
+        val pathDir = File.separator + appNamed() + "/" + appNamed() + " " + folderExtNamed
+        val storePath = Environment.getExternalStorageDirectory().absolutePath + pathDir
         val appDir = File(storePath)
         if (!appDir.exists()) {
             appDir.mkdir()
@@ -128,16 +155,15 @@ class FolderFileSaverPlugin(private val registrar: Registrar) : MethodCallHandle
     }
 
     private fun appNamed(): String {
-        val ctx = registrar.activeContext().applicationContext
         var aI: ApplicationInfo? = null
         try {
-            aI = ctx.packageManager.getApplicationInfo(ctx.packageName, 0)
+            aI = context.packageManager.getApplicationInfo(context.packageName, 0)
         } catch (err: PackageManager.NameNotFoundException) {
             err.printStackTrace()
         }
         val appName: String
         appName = if (aI != null) {
-            val cS = ctx.packageManager.getApplicationLabel(aI)
+            val cS = context.packageManager.getApplicationLabel(aI)
             StringBuilder(cS.length).append(cS).toString()
         } else {
             "Folder File Saver"
@@ -145,29 +171,10 @@ class FolderFileSaverPlugin(private val registrar: Registrar) : MethodCallHandle
         return appName
     }
 
-    private fun requirePermission() {
-        ActivityCompat.requestPermissions(registrar.activity(), arrayOf(permission), 0)
-    }
-
     private fun openSettingsPermission() {
-        val activity = registrar.activity()
         val intent = Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS, Uri.parse("package:${activity.packageName}"))
         intent.addCategory(Intent.CATEGORY_DEFAULT)
         intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK
         activity.startActivity(intent)
-    }
-
-    override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<out String>?, grantResults: IntArray?): Boolean {
-        var result = 0
-        val ac = registrar.activity()
-        val notAskAgain = ActivityCompat.shouldShowRequestPermissionRationale(ac, permission);
-        result = if (notAskAgain) {
-            1
-        } else 2
-        if (grantResults?.get(0) == PackageManager.PERMISSION_GRANTED) {
-            result = 0
-        }
-        mChannel.invokeMethod("resultPermission", result)
-        return grantResults!!.isEmpty() || grantResults[0] != PackageManager.PERMISSION_GRANTED
     }
 }