+++ example/lib/main.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:folder_file_saver/folder_file_saver.dart';
 import 'package:path_provider/path_provider.dart' as p;
 import 'package:dio/dio.dart';
+import 'package:permission_handler/permission_handler.dart';
 
 void main() => runApp(MyApp());
 
@@ -14,9 +15,10 @@ class _MyAppState extends State<MyApp> {
   String progress = "0";
   bool _isLoading = false;
   final urlVideo =
-          'http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/ElephantsDream.mp4',
+          'https://media.istockphoto.com/videos/lovely-puppy-labrador-running-to-the-camera-on-the-lawn-4k-video-id1073535242',
       urlImage =
           'https://images.unsplash.com/photo-1576039716094-066beef36943?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=634&q=80';
+
   Dio dio;
   @override
   void initState() {
@@ -26,24 +28,22 @@ class _MyAppState extends State<MyApp> {
 
   void _saveImage() async {
     try {
+      // get status permission
+      final status = await Permission.storage.status;
+
+      // check status permission
+      if (status.isDenied) {
+        // request permission
+        await Permission.storage.request();
+        return;
+      }
+
       setState(() {
         _isLoading = true;
       });
 
-      // 0 permission is PERMISSION_GRANTED
-      // 1 permission is PERMISSION_DENIED
-      // 2 permission is PERMISSION_DENIED wuth (don't ask again)
-      final resultPermission = await FolderFileSaver.requestPermission();
-
-      if (resultPermission == 2) {
-        // Do Something Info Here To User
-        // await FolderFileSaver.openSetting;
-      }
-
-      // Permission Granted
-      if (resultPermission == 0) {
-        await _doSaveImage();
-      }
+      // do save
+      await _doSaveImage();
     } catch (e) {
       print(e.toString());
     } finally {
@@ -55,31 +55,22 @@ class _MyAppState extends State<MyApp> {
 
   void _saveFolderFileExt() async {
     try {
+      // get status permission
+      final status = await Permission.storage.status;
+
+      // check status permission
+      if (status.isDenied) {
+        // request permission
+        await Permission.storage.request();
+        return;
+      }
+
       setState(() {
-        _isLoading = false;
+        _isLoading = true;
       });
-      // if you want check permission user
-      // use like that
-      // if return 0 permission is PERMISSION_GRANTED
-      // if return 1 permission is PERMISSION_IS_DENIED
-      // if return 2 permission is PERMISSION_IS_DENIED with click don't ask again
-      final resultPermission = await FolderFileSaver.requestPermission();
 
-      // 2 permission is PERMISSION_IS_DENIED with click don't ask again
-      if (resultPermission == 2) {
-        // Do Something Info Here To User
-        // await FolderFileSaver.openSetting;
-      }
-
-      // 1 permission is PERMISSION_IS_DENIED
-      if (resultPermission == 1) {
-        // Do Something Here
-      }
-
-      // 0 permission is PERMISSION_GRANTED
-      if (resultPermission == 0) {
-        await _doSave();
-      }
+      // do save
+      await _doSave();
     } catch (e) {
       print(e.toString());
     } finally {
@@ -89,9 +80,11 @@ class _MyAppState extends State<MyApp> {
     }
   }
 
+  // Don't forget to check
+  // device permission
   Future<void> _doSaveImage() async {
     final dir = await p.getTemporaryDirectory();
-    final pathImage = dir.path + ('example_image.png');
+    final pathImage = dir.path + ('/example_image.png');
     await dio.download(urlImage, pathImage, onReceiveProgress: (rec, total) {
       setState(() {
         progress = ((rec / total) * 100).toStringAsFixed(0) + "%";
@@ -102,20 +95,30 @@ class _MyAppState extends State<MyApp> {
     // cause default is return width = 0, height = 0
     // which will make it to get the original image
     // just write like this
-    final result = await FolderFileSaver.saveImage(pathImage: pathImage);
+    // remove originFile default = false
+    final result = await FolderFileSaver.saveImage(
+      pathImage: pathImage,
+      removeOriginFile: true,
+    );
     print(result);
   }
 
+  // Don't forget to check
+  // device permission
   Future<void> _doSave() async {
     final dir = await p.getTemporaryDirectory();
     // prepare the file and type extension that you want to download
-    final filePath = dir.path + ('example_video.mp4');
+    // remove originFile after success default = false
+    final filePath = dir.path + ('/example_video.mp4');
     await dio.download(urlVideo, filePath, onReceiveProgress: (rec, total) {
       setState(() {
         progress = ((rec / total) * 100).toStringAsFixed(0) + "%";
       });
     });
-    final result = await FolderFileSaver.saveFileToFolderExt(filePath);
+    final result = await FolderFileSaver.saveFileToFolderExt(
+      filePath,
+      removeOriginFile: true,
+    );
     print(result);
   }
 
@@ -125,7 +128,7 @@ class _MyAppState extends State<MyApp> {
     String result;
     final dir = await p.getTemporaryDirectory();
     // prepare the file and type extension that you want to download
-    final filePath = dir.path + ('example_video.mp4');
+    final filePath = dir.path + ('/example_video.mp4');
     try {
       await dio.download(urlVideo, filePath);
       result = await FolderFileSaver.saveFileToFolderExt(filePath);
@@ -141,7 +144,9 @@ class _MyAppState extends State<MyApp> {
       _isLoading = true;
     });
     // get your path from your device your device
-    final fileToCopy = '/storage/emulated/0/DCIM/Camera/20200102_202226.jpg';
+    // final fileToCopy = '/storage/emulated/0/DCIM/Camera/20200102_202226.jpg'; // example
+    // remove originFile default = false
+    final fileToCopy = '<local_path_from_your_device>';
     try {
       await FolderFileSaver.saveFileToFolderExt(fileToCopy);
     } catch (e) {
@@ -157,33 +162,36 @@ class _MyAppState extends State<MyApp> {
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
-          title: const Text('Folder File Saver'),
+          title: const Text('Folder File Saver Example'),
           centerTitle: true,
         ),
-        body: Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            RaisedButton(
-              onPressed: _isLoading ? null : _saveImage,
-              child: Text(_isLoading
-                  ? 'Downloading $progress'
-                  : 'Download Image and Resize'),
-            ),
-            RaisedButton(
-              onPressed: _isLoading ? null : _saveFolderFileExt,
-              child:
-                  Text(_isLoading ? 'Downloading $progress' : 'Download File'),
-            ),
-            RaisedButton(
-              onPressed: copyFileToNewFolder,
-              child: Text('Copy File to Folder'),
-            ),
-            RaisedButton(
-              onPressed: () async => await FolderFileSaver.openSetting,
-              child: Text('Open Setting App'),
-            ),
-          ],
+        body: Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.stretch,
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              ElevatedButton(
+                onPressed: _isLoading ? null : _saveImage,
+                child: Text(_isLoading
+                    ? 'Downloading $progress'
+                    : 'Download Image and Resize'),
+              ),
+              ElevatedButton(
+                onPressed: _isLoading ? null : _saveFolderFileExt,
+                child: Text(
+                    _isLoading ? 'Downloading $progress' : 'Download File'),
+              ),
+              ElevatedButton(
+                onPressed: copyFileToNewFolder,
+                child: Text('Copy File to Folder'),
+              ),
+              ElevatedButton(
+                onPressed: () async => await FolderFileSaver.openSetting,
+                child: Text('Open Setting App'),
+              ),
+            ],
+          ),
         ),
       ),
     );